Add preliminary support for node revisions, change node number to long

+++ src/main/java/fi/thl/termed/domain/DepthLimitedNodeTree.java
@@ -49,7 +49,7 @@ public class DepthLimitedNodeTree implements NodeTree {
   }
 
   @Override
-  public Integer getNumber() {
+  public Long getNumber() {
     return source.getNumber();
   }
